+++ packages/multimodal/agent-snapshot/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @multimodal/agent-snapshot
 
+## 0.1.1-beta.4
+
+### Patch Changes
+
+- @multimodal/agent@0.1.1-beta.4
+- @multimodal/agent-interface@0.1.1-beta.4
+
 ## 0.1.1-beta.3
 
 ### Patch Changes