+++ packages/multimodal/agent-tars-web-ui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @agent-tars/web-ui
 
+## 0.0.1-beta.6
+
 ## 0.0.1-beta.5
 
 ## 0.0.1-beta.4