+++ packages/multimodal/agent-tars/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @agent-tars/core
 
+## 0.0.1-beta.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @agent-infra/mcp-server-browser@1.1.6-beta.3
+  - @agent-infra/mcp-server-commands@1.1.6-beta.3
+  - @agent-infra/mcp-server-filesystem@1.1.6-beta.3
+  - @agent-infra/mcp-server-search@1.1.6-beta.3
+  - @multimodal/mcp-agent@0.1.1-beta.4
+  - @agent-infra/shared@0.0.1
+
 ## 0.0.1-beta.5
 
 ## 0.0.1-beta.4