+++ packages/multimodal/agent/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @multimodal/agent
 
+## 0.1.1-beta.4
+
+### Patch Changes
+
+- @multimodal/agent-interface@0.1.1-beta.4
+- @multimodal/llm-client@0.1.1-beta.4
+
 ## 0.1.1-beta.3
 
 ### Patch Changes