+++ packages/multimodal/llm-client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @multimodal/llm-client
 
+## 0.1.1-beta.4
+
 ## 0.1.1-beta.3
 
 ### Patch Changes